requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyTelegramBotAPI==4.22.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
